Let reading a written figure be its own, not a reader of ends
Working out what a clause states means knowing what number `5.00m` is, which
every carrier asks of the literals it meets. Those two functions were declared
inside `InvariantBound`, whose subject is where a rule placed an end — so a
check reading which types a classification depends on found it depending on a
reader of answers, through the one edge that is not one.

They depend on nothing of that type: an integer, a decimal, a negation, and the
scale a figure was written to. `NumericLiterals` owns them, the carrier calls
them there, and `InvariantBound` calls them like anybody else.

The same misplacement `LinearForm` and `Rel` were in, found the same way — by
walking what the classification actually computes with rather than what its
values hold.
